Add entry to lexicon

This function creates a new entry in the lexicon, i.e. a new empty entry skeleton is written to disk, in the lexicon/ directory, for the user to edit at will.

Usage

add_entry(
  lexadb,
  lexeme = NULL,
  gloss = NULL,
  part_of_speech = NULL,
  phon = NULL,
  morph_category = NULL,
  morph_type = NULL,
  definition = gloss,
  etymology = NULL,
  notes = NULL,
  homophone = NULL,
  edit = TRUE
)

Arguments

lexadb

A lexadb object (created with load_lexadb).

lexeme

The entry as a string.

gloss

The gloss as a string.

part_of_speech

The part of speech as a string.

phon

The phonetic transcription as a string.

morph_category

The morphosyntactic category as a string ("lexical" or "grammatical").

morph_type

The type of morpheme as a string.

definition

The definition of the entry as a string.

etymology

The etymology of the entry as a string.

notes

Further notes as a string.

homophone

The homophone numeric index.

edit

Open file for editing after creation (default is TRUE)

Value

Nothing. Used for its side effects