Show text or sentence with given id

It shows the text or sentence with the given id.

Usage

show_text(lexadb, text_id, sent_id = NULL)

Arguments

lexadb

A lexadb object (created with load_lexadb).

text_id

A string with the text id (the tx_ prefix and leading zeros can be omitted.)

sent_id

A string with the sentence id (the st_ prefix and leading zeros can be omitted.)

Value

A lexast object.

Examples

db_path <- system.file("extdata/albanian_lexadb", package = "lexa")
albanian <- load_lexadb(db_path)
#> ℹ Loading lexa database...

show_text(albanian, 1)
#> 
#> ── The North Wind and the Sun ──────────────────────────────────────────────────
#> 
#> ── st_000001 ──
#> 
#> Era e Veriut dhe Dielli po ziheshin se kush ishte më i fortë.
#> The North Wind and the Sun were disputing which was the stronger,
#> 
#> ── st_000002 ──
#> 
#> Kur aty kaloi një udhëtar që kishte veshur një pallto të ngrohtë.
#> when a traveller came along wrapped in a warm cloak.
#> 
#> ── st_000003 ──
#> 
#> Ata ranë dakord që kush do ta bënte udhëtarin të hiqte pallton më përpara do të
#> quhej më i forti.
#> They agreed that the one who first succeeded in making the traveller take his
#> cloak off first should be considered the stronger one.
#> 
#> ── st_000004 ──
#> 
#> Fill pas kësaj, Era e Veriut filloi të frynte me gjithë fuqinë e saj,
#> Then the North Wind blew as hard as he could,
#> 
#> ── st_000005 ──
#> 
#> por sa më shumë që frynte aq më shumë kapej udhëtari pas palltos së tij,
#> but the more he blew the more closely did the traveller fold his cloak around
#> him,
#> 
#> ── st_000006 ──
#> 
#> derisa më në fund Era e Veriut u dorëzua.
#> and at last the North Wind gave up the attempt.
#> 
#> ── st_000007 ──
#> 
#> Pastaj Dielli lëshoi rrezet e tij të ngrohta dhe menjëherë udhëtari e hoqi
#> pallton.
#> Then the Sun released his warm rays and immediately the traveller took off his
#> cloak.
#> 
#> ── st_000008 ──
#> 
#> E kështu Era e Veriut u detyrua të pranonte që Dielli ishte më i fortë se ajo.
#> And so the North Wind was forced to accept that the Sun was the stronger than
#> itself.
show_text(albanian, 1, 3)
#> 
#> ── Ata ranë dakord që kush do ta bënte udhëtarin të hiqte pallton më përpara do 
#> [ata ɽanɜ dakɔɽd | t̻ʃ̻ɜ kuʃ dɔ ta bɜnte uðɜtaɽin tɜ hit̻ʃ̻te palˠtɔn mɜ pɜɽpaɽa dɔ
#> tɜ t̻ʃ̻uhej mɜ i fɔɽti ‖]
#> 
#> ata         ra-në                  dakord      që     kush   do    t=a           
#> 3PL.M.NOM   AOR\fall-AOR.IND.3PL   agreement   that   who    Fut   Fut=3SG.ACC   
#> 
#> bën-te              udhëtar-in               të     hiq-te                       
#> make-IND.IMPF.3SG   traveller-M.SG.ACC.DEF   SBJV   SBJV\take_off-PST.SBJV.3SG   
#> 
#> pallt-on            më     përpara   do    të    quh-ej                   më     
#> coat-F.SG.ACC.DEF   more   soon      Fut   Fut   PASS\call-IMPF.IND.3SG   more   
#> 
#> i              fort-i                
#> M.SG.NOM.DEF   strong-M.SG.NOM.DEF   
#> 
#> 
#> ‘They agreed that the one who first succeeded in making the traveller take his
#> cloak off first should be considered the stronger one.’